Reorganize OpenLIFULib
The __init__ of OpenLIFULib has become long, so this is just
moving things into their own modules for easier maintenance.
This is part of work towards #84; the primary goal was to move
the OpenLIFUAlgorithmInputWidget definition into its own python module
