+++ OpenLIFULib/CMakeLists.txt
@@ -4,9 +4,15 @@ set(MODULE_NAME OpenLIFULib)
 #-----------------------------------------------------------------------------
 set(MODULE_PYTHON_SCRIPTS
   OpenLIFULib/__init__.py
-  OpenLIFULib/busycursor.py
+  OpenLIFULib/util.py
   OpenLIFULib/lazyimport.py
   OpenLIFULib/parameter_node_utils.py
+  OpenLIFULib/session.py
+  OpenLIFULib/transducer.py
+  OpenLIFULib/targets.py
+  OpenLIFULib/simulation.py
+  OpenLIFULib/algorithm_input_widget.py
+  OpenLIFULib/coordinate_system_utils.py
 )
 
 set(MODULE_PYTHON_RESOURCES
